+++ molo/surveys/forms.py
@@ -1,10 +1,16 @@
 import csv
+from collections import defaultdict
 
 from django import forms
 from django.contrib.auth import get_user_model
+from django.contrib.auth.models import Group
+from django.core.exceptions import ValidationError
+from django.forms.utils import ErrorList
 from django.utils.translation import ugettext_lazy as _
 
-from django.contrib.auth.models import Group
+from wagtail.wagtailadmin.forms import WagtailAdminPageForm
+
+from .blocks import SkipState, VALID_SKIP_SELECTORS
 
 
 class CSVGroupCreationForm(forms.ModelForm):
@@ -80,3 +86,123 @@ class CSVGroupCreationForm(forms.ModelForm):
 
         # Add users to the group
         group.user_set.add(*self.__initial_users)
+
+
+class BaseMoloSurveyForm(WagtailAdminPageForm):
+    def clean(self):
+        cleaned_data = super(BaseMoloSurveyForm, self).clean()
+        for form in self.formsets[self.form_field_name]:
+            self._clean_errors = defaultdict(lambda: defaultdict(list))
+            if form.is_valid():
+                data = form.cleaned_data
+                for i, logic in enumerate(data['skip_logic']):
+                    if logic.value['skip_logic'] == SkipState.SURVEY:
+                        survey = logic.value['survey']
+                        self.clean_survey(i, survey)
+                    if logic.value['skip_logic'] == SkipState.QUESTION:
+                        sort_order = logic.value['question'] - 1
+                        questions = (
+                            getattr(self.instance, self.form_field_name)
+                        )
+                        question = questions.get(sort_order=sort_order)
+                        self.clean_question(i, data['segment'], question)
+                if self.clean_errors:
+                    form._errors = self.clean_errors
+
+        return cleaned_data
+
+    def save(self, commit):
+        # Tidy up the skip logic when field cant have skip logic
+        for form in self.formsets[self.form_field_name]:
+            field_type = form.instance.field_type
+            if field_type not in VALID_SKIP_SELECTORS:
+                if field_type != 'checkboxes':
+                    form.instance.skip_logic = []
+                else:
+                    for skip_logic in form.instance.skip_logic:
+                        skip_logic.value['skip_logic'] = SkipState.NEXT
+                        skip_logic.value['question'] = None
+                        skip_logic.value['survey'] = None
+
+        return super(BaseMoloSurveyForm, self).save(commit)
+
+    def clean_question(self, position, *args):
+        self.clean_formset_field('question', position, *args)
+
+    def clean_survey(self, position, *args):
+        self.clean_formset_field('survey', position, *args)
+
+    def clean_formset_field(self, field, position, *args):
+        for method in getattr(self, field + '_clean_methods', []):
+            error = getattr(self, method)(*args)
+            if error:
+                self.add_stream_field_error(position, field, error)
+
+    def check_doesnt_loop_to_self(self, survey):
+        if survey and self.instance == survey:
+            return _('Cannot skip to self, please select a different survey.')
+
+    def add_stream_field_error(self, position, field, message):
+        self._clean_errors[position][field].append(message)
+
+    @property
+    def clean_errors(self):
+        if self._clean_errors.keys():
+            params = {
+                key: ErrorList(
+                    [ValidationError('Error in form', params=value)]
+                )
+                for key, value in self._clean_errors.items()
+            }
+            return {
+                'skip_logic': ErrorList([ValidationError(
+                    'Skip Logic Error',
+                    params=params,
+                )])
+            }
+
+
+class MoloSurveyForm(BaseMoloSurveyForm):
+    form_field_name = 'survey_form_fields'
+    survey_clean_methods = [
+        'check_doesnt_loop_to_self',
+        'check_doesnt_link_personalised_survey',
+    ]
+
+    def check_doesnt_link_to_peronsalised_survey(self, survey):
+        try:
+            segment = survey.personalisablesurvey.segment
+        except AttributeError:
+            pass
+        else:
+            # Can only link a survey without a segments
+            if segment:
+                return _('Cannot select a survey with a segment.')
+
+
+class PersonalisableMoloSurveyForm(BaseMoloSurveyForm):
+    form_field_name = 'personalisable_survey_form_fields'
+    survey_clean_methods = [
+        'check_doesnt_loop_to_self',
+        'check_survey_link_valid',
+    ]
+
+    question_clean_methods = [
+        'check_question_segment_ok',
+    ]
+
+    def check_question_segment_ok(self, current_segment, linked_question):
+        segment = linked_question.segment
+        # Cannot link from None to segment, but can link from segment to None
+        if (segment and not current_segment) or (segment != current_segment):
+            return _('Cannot link to a question with a different segment.')
+
+    def check_survey_link_valid(self, survey):
+        try:
+            segment = survey.personalisablesurvey.segment
+        except AttributeError:
+            pass
+        else:
+            # Can only link a survey without segments or the same segment
+            if segment and segment != self.instance.segment:
+                return _('Cannot select a survey with a different segment.')