+++ molo/surveys/tests/test_views.py
@@ -1,14 +1,18 @@
+from bs4 import BeautifulSoup
 from django.contrib.auth import get_user_model
 from django.core.urlresolvers import reverse
 from django.test import TestCase
 from django.test.client import Client
-
-from molo.core.models import SiteLanguageRelation, Main, Languages
+from django.utils.text import slugify
+from molo.core.models import Languages, Main, SiteLanguageRelation
 from molo.core.tests.base import MoloTestCaseMixin
-from molo.surveys.models import (MoloSurveyPage, MoloSurveyFormField,
-                                 SurveysIndexPage)
+from molo.surveys.models import (
+    MoloSurveyFormField,
+    MoloSurveyPage,
+    SurveysIndexPage,
+)
 
-from bs4 import BeautifulSoup
+from .utils import skip_logic_data
 
 User = get_user_model()
 
@@ -606,3 +610,195 @@ class TestDeleteButtonRemoved(TestCase, MoloTestCaseMixin):
                          'class="shortcut">Delete</a></li>'
                          .format(str(surveys_index_page.pk)))
         self.assertNotContains(response, delete_button, html=True)
+
+
+class TestSkipLogicSurveyView(TestCase, MoloTestCaseMixin):
+    def setUp(self):
+        self.mk_main()
+        self.main = Main.objects.all().first()
+        self.language_setting = Languages.objects.create(
+            site_id=self.main.get_site().pk)
+        self.english = SiteLanguageRelation.objects.create(
+            language_setting=self.language_setting,
+            locale='en',
+            is_active=True)
+
+        self.molo_survey_page = self.new_survey('Test Survey')
+        self.another_molo_survey_page = self.new_survey('Another Test Survey')
+
+        self.last_molo_survey_form_field = MoloSurveyFormField.objects.create(
+            page=self.molo_survey_page,
+            sort_order=3,
+            label='Your favourite actor',
+            field_type='singleline',
+            required=True
+        )
+
+        self.choices = ['next', 'end', 'survey', 'question']
+        self.skip_logic_form_field = MoloSurveyFormField.objects.create(
+            page=self.molo_survey_page,
+            sort_order=1,
+            label='Where should we go',
+            field_type='dropdown',
+            skip_logic=skip_logic_data(
+                self.choices,
+                self.choices,
+                survey=self.another_molo_survey_page,
+                question=self.last_molo_survey_form_field,
+            ),
+            required=True
+        )
+
+        self.molo_survey_form_field = MoloSurveyFormField.objects.create(
+            page=self.molo_survey_page,
+            sort_order=2,
+            label='Your favourite animal',
+            field_type='singleline',
+            required=True
+        )
+
+        self.another_molo_survey_form_field = (
+            MoloSurveyFormField.objects.create(
+                page=self.another_molo_survey_page,
+                sort_order=1,
+                label='Your favourite actress',
+                field_type='singleline',
+                required=True
+            )
+        )
+
+    def new_survey(self, name):
+        survey = MoloSurveyPage(
+            title=name, slug=slugify(name),
+            intro='Introduction to {}...'.format(name),
+            thank_you_text='Thank you for taking the {}'.format(name),
+            submit_text='survey submission text for {}'.format(name),
+            allow_anonymous_submissions=True,
+        )
+        self.section_index.add_child(instance=survey)
+        survey.save_revision().publish()
+        return survey
+
+    def assertSurveyAndQuestions(self, response, survey, questions):
+        self.assertContains(response, survey.title)
+        self.assertContains(response, survey.intro)
+        for question in questions:
+            self.assertContains(response, question.label)
+            self.assertContains(response, question.label)
+
+    def test_skip_logic_next_question(self):
+        response = self.client.get(self.molo_survey_page.url)
+
+        self.assertSurveyAndQuestions(
+            response,
+            self.molo_survey_page,
+            [self.skip_logic_form_field]
+        )
+        self.assertNotContains(
+            response,
+            self.last_molo_survey_form_field.label
+        )
+        self.assertNotContains(response, self.molo_survey_form_field.label)
+        self.assertContains(response, 'Next Question')
+
+        response = self.client.post(self.molo_survey_page.url + '?p=2', {
+            self.skip_logic_form_field.clean_name: self.choices[0],
+        })
+
+        self.assertSurveyAndQuestions(
+            response,
+            self.molo_survey_page,
+            [self.molo_survey_form_field, self.last_molo_survey_form_field]
+        )
+        self.assertNotContains(response, self.skip_logic_form_field.label)
+        self.assertContains(response, self.molo_survey_page.submit_text)
+
+        response = self.client.post(self.molo_survey_page.url + '?p=3', {
+            self.molo_survey_form_field.clean_name: 'python',
+            self.last_molo_survey_form_field.clean_name: 'Steven Seagal ;)',
+        }, follow=True)
+
+        self.assertContains(response, self.molo_survey_page.thank_you_text)
+
+    def test_skip_logic_to_end(self):
+        response = self.client.get(self.molo_survey_page.url)
+
+        self.assertSurveyAndQuestions(
+            response,
+            self.molo_survey_page,
+            [self.skip_logic_form_field]
+        )
+        self.assertNotContains(
+            response,
+            self.last_molo_survey_form_field.label,
+        )
+        self.assertNotContains(response, self.molo_survey_form_field.label)
+        self.assertContains(response, 'Next Question')
+
+        response = self.client.post(self.molo_survey_page.url + '?p=2', {
+            self.skip_logic_form_field.clean_name: self.choices[1],
+        }, follow=True)
+
+        # Should end the survey and not complain about required
+        # field for the last field
+
+        self.assertContains(response, self.molo_survey_page.title)
+        self.assertNotContains(response, self.molo_survey_form_field.label)
+        self.assertNotContains(
+            response,
+            self.last_molo_survey_form_field.label
+        )
+        self.assertNotContains(response, self.molo_survey_page.submit_text)
+        self.assertContains(response, self.molo_survey_page.thank_you_text)
+
+    def test_skip_logic_to_another_survey(self):
+        response = self.client.get(self.molo_survey_page.url)
+
+        self.assertSurveyAndQuestions(
+            response,
+            self.molo_survey_page,
+            [self.skip_logic_form_field]
+        )
+        self.assertNotContains(
+            response,
+            self.last_molo_survey_form_field.label
+        )
+        self.assertNotContains(response, self.molo_survey_form_field.label)
+        self.assertContains(response, 'Next Question')
+
+        response = self.client.post(self.molo_survey_page.url + '?p=2', {
+            self.skip_logic_form_field.clean_name: self.choices[2],
+        }, follow=True)
+
+        # Should end the survey and progress to the new survey
+        self.assertSurveyAndQuestions(
+            response,
+            self.another_molo_survey_page,
+            [self.another_molo_survey_form_field],
+        )
+
+    def test_skip_logic_to_another_question(self):
+        response = self.client.get(self.molo_survey_page.url)
+
+        self.assertSurveyAndQuestions(
+            response,
+            self.molo_survey_page,
+            [self.skip_logic_form_field]
+        )
+        self.assertNotContains(
+            response,
+            self.last_molo_survey_form_field.label,
+        )
+        self.assertNotContains(response, self.molo_survey_form_field.label)
+        self.assertContains(response, 'Next Question')
+
+        response = self.client.post(self.molo_survey_page.url + '?p=2', {
+            self.skip_logic_form_field.clean_name: self.choices[3],
+        }, follow=True)
+
+        # Should end the survey and progress to the new survey
+        self.assertSurveyAndQuestions(
+            response,
+            self.molo_survey_page,
+            [self.last_molo_survey_form_field],
+        )